% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relativepositions.R
\name{distRanges}
\alias{distRanges}
\title{distRanges}
\usage{
distRanges(ranges1, ranges2, strand = 0)
}
\arguments{
\item{ranges1}{GRanges object}

\item{ranges2}{GRanges object}

\item{strand}{either 1,2, or 0 if dist is relative to strand of ranges1, ranges2, or neither}
}
\value{
returns either subset of nucs that are +1 or annotated version of sites
}
\description{
distRanges
}
\seealso{
\code{\link{get_m1_nucs}}  \code{\link{get_p1_nucs}}  \code{\link{get_dist_between_calls}}
}

