% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relativepositions.R
\name{get_dist_between_calls}
\alias{get_dist_between_calls}
\title{get_dist_between_calls}
\usage{
get_dist_between_calls(ranges, max_dist = 1000)
}
\arguments{
\item{ranges}{GRanges object}

\item{max_dist}{to consider}
}
\value{
returns tabulated distance between calls
}
\description{
get_dist_between_calls
}
\seealso{
\code{\link{distRanges}}
}

