% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relativepositions.R
\name{getstrand}
\alias{getstrand}
\title{getstrand}
\usage{
getstrand(ranges)
}
\arguments{
\item{ranges}{GRanges object}
}
\value{
returns strand as character (rather than factor Rle)
}
\description{
getstrand
}
\seealso{
\code{\link{shiftRanges}}
}

