% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{mean_smooth}
\alias{mean_smooth}
\title{mean_smooth
Smooth using a flat window}
\usage{
mean_smooth(x, window = 11)
}
\arguments{
\item{x}{vector of values}

\item{window}{size of window for smoothing}
}
\value{
returns vector of size x that has been smoothed.  For values at edge, mean is for window included in x.
(i.e. for fist value, smoothed valued is mean of that value and the window/2 values after)
}
\description{
mean_smooth
Smooth using a flat window
}

