% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NucleoATACR-package.R
\name{NucleoATACR}
\alias{NucleoATACR}
\title{NucleoATACR}
\description{
Package with some functions for reading and manipulating results from NucleoATAC python package.
}
\author{
Alicia Schep
}
\seealso{
\code{\link{readNucs}}, \code{\link{readNFRs}}, \code{\link{plotV}}
}

