% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vplot.R
\name{order_of_magnitude}
\alias{order_of_magnitude}
\title{Determine order of magnitude}
\usage{
order_of_magnitude(x)
}
\arguments{
\item{x}{number}
}
\value{
numeric
}
\description{
Determine order of magnitude
}

