% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vplot.R
\name{plotV}
\alias{plotV}
\title{plotV}
\usage{
plotV(X, xlabel = "Center position relative to dyad (bp)",
  ylabel = "Fragment size (bp)", guide = "Density", name = NA,
  palette = "BuPu", limits = NA, xbreaks = NA, ybreaks = NA)
}
\arguments{
\item{X}{vplot matrix}

\item{xlabel}{label for x axis}

\item{ylabel}{label for y axis}

\item{guide}{label for legend}

\item{name}{title}

\item{palette}{color palette}

\item{limits}{limits for color}

\item{xbreaks}{where to include breaks for x axis}

\item{ybreaks}{wehre to include breaks for y axis}
}
\value{
returns ggplot object
}
\description{
plotV
}
\seealso{
\code{\link{read_vplot}} \code{\link{vplot_theme}}
}

