% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vplot.R
\name{pretty_scale_format}
\alias{pretty_scale_format}
\title{Format numbers in pretty manner}
\usage{
pretty_scale_format(l)
}
\arguments{
\item{l}{list of numbers}
}
\value{
expression with formatted numbers
}
\description{
Format numbers in pretty manner
}
\seealso{
\code{\link{pretty_scientific}}
}

