% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vplot.R
\name{pretty_scientific}
\alias{pretty_scientific}
\title{Format numbers in scientific notation}
\usage{
pretty_scientific(l)
}
\arguments{
\item{l}{list of numbers}
}
\value{
expression with formatted numbers
}
\description{
Format numbers in scientific notation
}
\seealso{
\code{\link{pretty_scale_format}}
}

