% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vplot.R
\name{read_vplot}
\alias{read_vplot}
\title{read_vplot}
\usage{
read_vplot(file, xlims = NA)
}
\arguments{
\item{file}{.VMat file}

\item{xlims}{limits for position axis}
}
\value{
returns matrix with vplot informatiom
}
\description{
read_vplot
}
\seealso{
\code{\link{plotV}}
}

