% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readoutputs.R
\name{readBedgraph}
\alias{readBedgraph}
\title{readBedgraph}
\usage{
readBedgraph(bgfile, chrom, start, end, empty = NA)
}
\arguments{
\item{bgfile}{tabix-indexed bedgraph file}

\item{chrom}{chromosome}

\item{start}{start (0-based)}

\item{end}{end (1-based)}

\item{empty}{value for positions not in bedgraph to take (default NA)}
}
\value{
returns vector of bedgraph values for interval
}
\description{
readBedgraph
}
\seealso{
\code{\link{readNucs}}  \code{\link{readNFRs}}
}

