% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readoutputs.R
\name{readNFRs}
\alias{readNFRs}
\title{readNFRs}
\usage{
readNFRs(nucfile, format = "GRanges")
}
\arguments{
\item{format}{desired object type, either data.frame or GRanges}

\item{nfrfile}{file of nucleosome positions from NucleoATAC}
}
\value{
returns object of type specified by format, either data.frame or GRanges
}
\description{
readNFRs
}
\seealso{
\code{\link{readNucs}}
}

