% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readoutputs.R
\name{readNucs}
\alias{readNucs}
\title{readNucs}
\usage{
readNucs(nucfile, out = "GRanges")
}
\arguments{
\item{nucfile}{file of nucleosome positions from NucleoATAC}

\item{out}{desired object type, either data.frame or GRanges}
}
\value{
returns object of type specified by format, either data.frame or GRanges
}
\description{
readNucs
}
\seealso{
\code{\link{readNFRs}}
}

