% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relativepositions.R
\name{shiftRanges}
\alias{shiftRanges}
\title{shiftRanges}
\usage{
shiftRanges(ranges, shift = 0)
}
\arguments{
\item{ranges}{GRanges object}

\item{shift}{amount by which to shift ranges (negative means shift upstream, positive downstream)}
}
\value{
returns GRanges object for which each range has been shifted appropriately
}
\description{
shiftRanges
}
\seealso{
\code{\link{get_m1_nuc}} \code{\link{get_p1_nuc}} \code{\link{distRanges}}
}

