% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{tabulate2}
\alias{tabulate2}
\title{tabulate2
Modified version of tabulate so that it gives counts in a given span and can handle negative integers}
\usage{
tabulate2(x, min_val, max_val)
}
\arguments{
\item{x}{vector of values}

\item{min_val}{minimum value}

\item{max_val}{maximum value}
}
\value{
returns vector of length max_val - min_val + 1 with values representing counts for values between min_val and max_val
}
\description{
tabulate2
Modified version of tabulate so that it gives counts in a given span and can handle negative integers
}

