% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vplot.R
\name{vplot_theme}
\alias{vplot_theme}
\title{vplot_theme}
\usage{
vplot_theme(base_size = 7, base_family = "Helvetica")
}
\arguments{
\item{base_size}{size for text}

\item{base_family}{font family}
}
\value{
returns theme for ggplot
}
\description{
vplot_theme
}
\seealso{
\code{\link{plotV}}
}

