% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relativepositions.R
\name{get_p1_nuc}
\alias{get_p1_nuc}
\title{get_p1_nuc}
\usage{
get_p1_nuc(nucs.ranges, sites, annotate_sites = F, max_dist = 250,
  shift = 0)
}
\arguments{
\item{nucs.ranges}{GRanges object with nucleosome positions}

\item{sites}{GRanges object with sites to get +1 nucs from.}

\item{annotate_sites}{what to return-- sites with +1 nuc annotation or +1 nucs?}

\item{max_dist}{what is maximum distance to consider}

\item{shift}{integer shift to apply to sites before finding -1}
}
\value{
returns either subset of nucs that are +1 or annotated version of sites
}
\description{
get_p1_nuc
}
\seealso{
\code{\link{get_m1_nuc}}
}

